Making a commit on working files.
The aesdsocket.c open a socket, receives input from an external source, logs
the incoming connection's IP (thus far I can only confirm IPv6 writes back
due to my particular testing environment but the IPv4 mechanisms are all
there), writes to a file, reads the file, sends content back to the calling
program before severing connection.

I still have to remove the temporary file, log the when the connection
disconnects, remove all of the dead DEBUGGING code, and ensure
'cross-compilation' is added to the Makefile.

As usual, it's a mess in there!  But it's working!  :-)

I have not committed the client software to test the incoming
connection.  That may occur in a later commit after '$scp -r <user@ip> .' from
the remote server.

	new file:   server/Makefile
	new file:   server/aesdsocket.c